Попытался пофиксить ошибку

+++ Красников_задания_асmр.ru/Program.cs
@@ -11,7 +11,7 @@ namespace Красников_задания_асmр.ru
         static void Main(string[] args)
         {
 
-            Console.WriteLine(Convert.ToInt32(Console.ReadLine()) + Convert.ToInt32(Console.ReadLine()));
+            Console.WriteLine(Convert.ToInt64(Console.ReadLine()) + Convert.ToInt64(Console.ReadLine()));
         }
     }
 }
